Rename the project package, so the archetype creation works.
