+++ archetypes/gwtp-basic/pom.xml
@@ -3,9 +3,9 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.arcbees.project</groupId>
-    <version>1.0-SNAPSHOT</version>
+    <groupId>com.arcbees.archetypes</groupId>
     <artifactId>gwtp-basic</artifactId>
+    <version>1.0-SNAPSHOT</version>
     <packaging>war</packaging>
     <name>GWTP Basic</name>
     <description>Basic GWTP application</description>
@@ -88,7 +88,7 @@
 
                     <runTarget>Project.html</runTarget>
                     <modules>
-                        <module>com.arcbees.project.Project</module>
+                        <module>org.arcbees.project.Project</module>
                     </modules>
                 </configuration>
                 <executions>
