+++ archetypes/gwtp-basic/src/main/java/org/arcbees/project/client/application/ApplicationView.ui.xml
