+++ archetypes/gwtp-basic/src/main/java/org/arcbees/project/client/application/home/HomeModule.java
@@ -14,19 +14,14 @@
  * the License.
  */
 
-package com.arcbees.project.client.application.home;
+package org.arcbees.project.client.application.home;
 
-import com.google.gwt.uibinder.client.UiBinder;
-import com.google.gwt.user.client.ui.Widget;
-import com.google.inject.Inject;
-import com.gwtplatform.mvp.client.ViewImpl;
+import com.gwtplatform.mvp.client.gin.AbstractPresenterModule;
 
-public class HomePageView extends ViewImpl implements HomePagePresenter.MyView {
-    public interface Binder extends UiBinder<Widget, HomePageView> {
-    }
-
-    @Inject
-    public HomePageView(final Binder uiBinder) {
-        initWidget(uiBinder.createAndBindUi(this));
+public class HomeModule extends AbstractPresenterModule {
+    @Override
+    protected void configure() {
+        bindPresenter(HomePagePresenter.class, HomePagePresenter.MyView.class, HomePageView.class,
+                HomePagePresenter.MyProxy.class);
     }
 }
