+++ archetypes/gwtp-basic/src/main/java/org/arcbees/project/client/application/home/HomePagePresenter.java
@@ -14,31 +14,30 @@
  * the License.
  */
 
-package com.arcbees.project.client.application;
+package org.arcbees.project.client.application.home;
 
-import com.google.gwt.event.shared.GwtEvent.Type;
+import org.arcbees.project.client.application.ApplicationPresenter;
+import org.arcbees.project.client.place.NameTokens;
+
 import com.google.inject.Inject;
 import com.google.web.bindery.event.shared.EventBus;
 import com.gwtplatform.mvp.client.Presenter;
 import com.gwtplatform.mvp.client.View;
-import com.gwtplatform.mvp.client.annotations.ContentSlot;
+import com.gwtplatform.mvp.client.annotations.NameToken;
 import com.gwtplatform.mvp.client.annotations.ProxyStandard;
-import com.gwtplatform.mvp.client.proxy.Proxy;
-import com.gwtplatform.mvp.client.proxy.RevealContentHandler;
+import com.gwtplatform.mvp.client.proxy.ProxyPlace;
 
-public class ApplicationPresenter extends Presenter<ApplicationPresenter.MyView, ApplicationPresenter.MyProxy> {
+public class HomePagePresenter extends Presenter<HomePagePresenter.MyView, HomePagePresenter.MyProxy> {
     public interface MyView extends View {
     }
 
-    @ContentSlot
-    public static final Type<RevealContentHandler<?>> TYPE_SetMainContent = new Type<RevealContentHandler<?>>();
-
     @ProxyStandard
-    public interface MyProxy extends Proxy<ApplicationPresenter> {
+    @NameToken(NameTokens.home)
+    public interface MyProxy extends ProxyPlace<HomePagePresenter> {
     }
 
     @Inject
-    public ApplicationPresenter(final EventBus eventBus, final MyView view, final MyProxy proxy) {
-        super(eventBus, view, proxy, RevealType.Root);
+    public HomePagePresenter(final EventBus eventBus, final MyView view, final MyProxy proxy) {
+        super(eventBus, view, proxy, ApplicationPresenter.TYPE_SetMainContent);
     }
 }
