+++ archetypes/gwtp-basic/src/main/java/org/arcbees/project/client/application/home/HomePageView.ui.xml
